Revert "Rename fuldecay to fromdecay."
This reverts commit 02549c5f31750d16c43768ffcb8b2f4a0a8e0ee8.
