+++ tests/fulldecay/__init__.py
@@ -1,0 +1,4 @@
+import pytest
+
+# This makes it so that assert errors are more helpful for e.g., the check_norm helper function
+pytest.register_assert_rewrite("fulldecay.test_fulldecay")
