+++ tests/test_nbody_decay.py
